Allow passing firstpass_image to txt2img API
This change allows passing in a base64 encoded image string to the
`/sdapi/v1/txt2img` endpoint via the `firstpass_image` parameter.

Can be useful when applications utilizing the API generate an initial
image and then need that image upscaled using HiRes Fix - by replaying
the request alongside the `firstpass_image` parameter with the base64
image returned by the original request, it bypasses the need to
regenerate the original image since it has already been passed in.

At least, that seems to be the original intent of this parameter (and
works great with this change).